python-client/requirements.txt
grpcio==1.67.0
grpcio-tools==1.67.0
protobuf==5.28.2